+++ source/processus/Server.h
@@ -19,7 +19,7 @@ public:
 	Server(double nTicks, std::string tick_unit);
 	virtual ~Server();
 
-	void update(double tic);
+	void update(unsigned tic);
 	double workValphen(double val = 0, bool set = false) { return NULL; }
 	double etatCurr(double val = 0, bool set = false) { return NULL; }
